Refactor of the Upload component
