+++ tests/d.Upload.js
@@ -16,7 +16,7 @@
 const assert = require('chai').assert;
 const expect = require('chai').expect;
 const should = require('chai').should();
-const Upload = require('../src/js/components/core/Upload.js');
+const Upload = require('../src/js/components/core/UploadMethods.js');
 const CheckStore = require('../src/js/stores/CheckStore.js')
 const path = require('path-extra');
 const api = window.ModuleApi;
